+++ changelog.md
@@ -2,6 +2,14 @@
 
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
+## 0.8.7(2023-02-09)
+
+- Feat: virtual fields can't be written
+- Feat: virtual fields without function can be read (for DB computed stuff)
+- Test: added tests for virtual dataFields
+- Fix: Issue with undefined dates
+- Chore: Cleaned some dead code
+
 ## 0.8.6(2023-02-08)
 
 - Feat: Node validations can also use the parent object of the mutation